requirements.txt
Babel==2.12.1
matplotlib==3.8.0
numpy==1.25.2
pandas==2.1.4
seaborn==0.13.0
streamlit==1.26.0
Pillow==10.2.0